requirements.txt
paho-mqtt>=1.6.0
aiohttp>=3.8.0
voluptuous>=0.13.0
